cambiato tipo somma da int a double

+++ gdb/gdb1.c
@@ -13,7 +13,7 @@
 
 int main()
 {
-	int somma = 0;
+	double somma = 0;
 	int val;
 	int numeri = 0;
 	while (scanf("%d\n",&val) != EOF) {
